# fix performance regression. Do not escape request path in echo.ServeH…